% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrectionPlots.R
\name{CorrectionPlots}
\alias{CorrectionPlots}
\title{CorrectionPlots}
\usage{
CorrectionPlots(
  data,
  AllDiatomsNames,
  Allcolor,
  MaxAge,
  NonNegative = TRUE,
  intervallBy = 100,
  MinAgeIntervall = 1
)
}
\arguments{
\item{data}{List of data generates by the Ordination function.}

\item{AllDiatomsNames}{All Diatom names from the Ordination function.}

\item{Allcolor}{The color from the Ordination function.}

\item{MaxAge}{MayAge from the Ordination function.}

\item{NonNegative}{Creates Positive Values after Loess calculation.}

\item{intervallBy}{Intervalls by to interpolate to.}

\item{MinAgeIntervall}{Minimal Intervall to interpolate to "Data Resolution".}
}
\value{
nothing.
}
\description{
Plots Ordinations Plots for the Pinoeer correction
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kroeger
}
