% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CutOutPionierPhase.R
\name{CutOutPionierPhase}
\alias{CutOutPionierPhase}
\title{CutOutPionierPhase}
\usage{
CutOutPionierPhase(
  data,
  intervallBy = 100,
  allLoessSpans = 0.3,
  minimumRowsAfterInterpolating = 12,
  method = "bray",
  AgeBuffer = 1000,
  AgeBuffer2 = 500
)
}
\arguments{
\item{data}{List of data generates by the Multivar function.}

\item{intervallBy}{Intervals by to interpolate to.}

\item{allLoessSpans}{span value for all Loess calculations made by Multivar.}

\item{minimumRowsAfterInterpolating}{Value for the minimum rows after filtering.}

\item{method}{The method for the dissimilarity calculation.}

\item{AgeBuffer}{How much age get Cut away.}

\item{AgeBuffer2}{How much age get Cut away as the second value.}
}
\value{
Returns the same List but with new added parameters.
}
\description{
Cut out the Pionier Phase of the Lake data.
\cr The new stating value is the first Value that passes the T.test that for two samples with a P_value <=0.05.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
\cr Comma numbers are rounded up.
}
\author{
Tim Kroeger
}
