% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSingalAfterTable.R
\name{DataSignalAfterTable}
\alias{DataSignalAfterTable}
\title{DataSignalAfterTable}
\usage{
DataSignalAfterTable(
  DataAllInOneTabel,
  BasicAsyncTable,
  ValueCantBeSamlerThanZero = FALSE
)
}
\arguments{
\item{DataAllInOneTabel}{Tabel of all data where Rows are the Time and columns are the variabels.}

\item{BasicAsyncTable}{An AsyncTable based on the mean value.}

\item{ValueCantBeSamlerThanZero}{If true, values smaler than 0 become 0.}
}
\value{
nothing.
}
\description{
Calculates the vector orientation of a data Table.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kroeger
}
