% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterAfterLakeData.R
\name{FilterAfterLakeData}
\alias{FilterAfterLakeData}
\title{FilterAfterLakeData}
\usage{
FilterAfterLakeData(data, distanceToCoast = 10)
}
\arguments{
\item{data}{List of data generates by the Multivar function.}

\item{distanceToCoast}{The minimum wanted distance to coast in km.}
}
\value{
Returns the same List but with less parameters.
}
\description{
Filters data based on the Lakedata Table
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
\cr Comma numbers are rounded up.
}
\author{
Tim Kroeger
}
