% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDicriptionCSV.R
\name{GetDiscriptionCSV}
\alias{GetDiscriptionCSV}
\title{GetDiscriptionCSV}
\usage{
GetDiscriptionCSV(data, saveAsCSV = FALSE)
}
\arguments{
\item{data}{List of data generates by the MultiExcelLoader function.}

\item{saveAsCSV}{Bool indicating if the data should be exported a CSV in the used working directory.}
}
\value{
Save CSV in workspace.
}
\description{
Get a short Discription from all Cores and save them as csv.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kroeger
}
