% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuessLoess.R
\name{GuessLoess}
\alias{GuessLoess}
\title{GuessLoess}
\usage{
GuessLoess(intervall, values, overspan = 20)
}
\arguments{
\item{intervall}{Intervall for the x-Achxes.}

\item{values}{values for they-achxes.}

\item{overspan}{Percent added to the minmal possible span.}
}
\value{
Returns the same List but with new added parameters.
}
\description{
Guess the best span for any function. Resolution for span = 0.01
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
\cr Comma numbers are rounded up.
}
\author{
Tim Kroeger.
}
