% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportDatabase.R
\name{ImportDatabase}
\alias{ImportDatabase}
\title{ImportDatabase}
\usage{
ImportDatabase(directory = NULL)
}
\arguments{
\item{directory}{specifyed directory.}
}
\value{
Downloaded data is stored in the data Folder.
}
\description{
Import data form the AWI server.
}
\details{
AWI VPN connection is needed.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kroeger
}
