% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverseSimpsion.R
\name{inverseSimpsion}
\alias{inverseSimpsion}
\title{inverseSimpsion}
\usage{
inverseSimpsion(
  data,
  intervallBy = 100,
  allLoessSpans = 0.8,
  minimumRowsAfterInterpolating = 0
)
}
\arguments{
\item{data}{List of data generates by the Multivar function.}

\item{intervallBy}{Intervalls by to interpolate to.}

\item{allLoessSpans}{span value for all Loess calculations made by Multivar.}

\item{minimumRowsAfterInterpolating}{Value for the minimum rows after filtering.}
}
\value{
Returns the same List but with new added parameters.
}
\description{
Calculates the inverse Simpsion for the multivar function.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
\cr Comma numbers are rounded up.
}
\author{
Tim Kroeger
}
