% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataFilter.R
\name{MetadataFilter}
\alias{MetadataFilter}
\title{MetadataFilter}
\usage{
MetadataFilter(
  data,
  NAignore = TRUE,
  Filter1 = c(name = "metdataName", criteria = c("greater", "smaller", "TRUE",
    "FALSE"), value = NULL),
  Filter2 = c(name = "metdataName", criteria = c("greater", "smaller", "TRUE",
    "FALSE"), value = NULL),
  Filter3 = c(name = "metdataName", criteria = c("greater", "smaller", "TRUE",
    "FALSE"), value = NULL),
  Filter4 = c(name = "metdataName", criteria = c("greater", "smaller", "TRUE",
    "FALSE"), value = NULL),
  Filter5 = c(name = "metdataName", criteria = c("greater", "smaller", "TRUE",
    "FALSE"), value = NULL)
)
}
\arguments{
\item{data}{List of data generates by the Multivar function.}

\item{NAignore}{If TRUE, value that are NA don't get filtered.}

\item{Filter1}{\cr\cr name = name of the metadata. Get this from the function ShowMetaDataFilters.
\cr\cr criteria = Can be "greater","smaller","TRUE" or"FALSE".
\cr\cr value = The filter value when greater or smaller.}

\item{Filter2}{Values for Filter 2.}

\item{Filter3}{Values for Filter 3.}

\item{Filter4}{Values for Filter 4.}

\item{Filter5}{Values for Filter 5.}
}
\value{
.Returns the same List but with filtert parameters
}
\description{
Filters data based on the Lakedata Table and other Metadata. To see with data fits please use the ShowMetaDataFilters function.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
\cr Comma numbers are rounded up.
}
\author{
Tim Kroeger
}
