% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiExcelLoader.R
\name{MultiExcelLoader}
\alias{MultiExcelLoader}
\title{MultiExcelLoader}
\usage{
MultiExcelLoader()
}
\value{
Retruns a List of all excel sheets.
}
\description{
Loads multiple Excels from a sub folder named data in your working directory.
}
\details{
When no Sub directory is give, this function will create the folder.
\cr When age data is given, all depth will be given a corresponding age.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kroeger
}
