% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivar.R
\name{Multivar}
\alias{Multivar}
\title{Multivar}
\usage{
Multivar(data)
}
\arguments{
\item{data}{List of data generates by the MultiExcelLoader function.}
}
\value{
Returns the same List but with new added parameters.
}
\description{
Calculates several parameters from deatom data.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kroeger
}
