% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrdinationPlots.R
\name{Ordination}
\alias{Ordination}
\title{Ordination}
\usage{
Ordination(data)
}
\arguments{
\item{data}{List of data generates by the Multivar function.}
}
\value{
nothing.
}
\description{
Plots Ordinations Plots.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kroeger
}
