% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateofChange.R
\name{rateofChange}
\alias{rateofChange}
\title{rateofChange}
\usage{
rateofChange(
  data,
  intervallBy = 100,
  minimumRowsAfterInterpolating = 0,
  method = "bray",
  MinAgeIntervall = 1,
  NonNegative = TRUE,
  Importname = "",
  Exportname = ""
)
}
\arguments{
\item{data}{List of data generates by the Multivar function.}

\item{intervallBy}{Intervalls by to interpolate to.}

\item{minimumRowsAfterInterpolating}{alue for the minimum rows after filtering.}

\item{method}{Method for calculation Dissimilarity Indices for Community Ecologists.}

\item{MinAgeIntervall}{Minimal Intervall to interpolate to "Data Resolution".}

\item{NonNegative}{Creates Positive Values after Loess calculation.}

\item{Importname}{importname after data$Diatom$DiatomNames$}

\item{Exportname}{data$Diatom$DiatomNames$}
}
\value{
Returns the same List but with new added parameters.
}
\description{
Calculates the Rate of change for the multivar function.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
\cr Comma numbers are rounded up.
}
\author{
Tim Kroeger
}
