% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowFilter.R
\name{RowFilter}
\alias{RowFilter}
\title{RowFilter}
\usage{
RowFilter(data, NoCDV = NULL, FilterByAlphaDiversity = F)
}
\arguments{
\item{data}{List of data generates by the MultiExcelLoader function.}

\item{NoCDV}{Custom number for numbers of counted diatom valves where the data will be filtered by.}

\item{FilterByAlphaDiversity}{If TRUE the data will be filtered by Alpha diversity instead of NoCDV.}
}
\value{
Returns the same List but with new added parameters.
}
\description{
Filters row by Total numbers of counted diatom valves.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kroeger
}
