% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveOutput.R
\name{SaveOutput}
\alias{SaveOutput}
\title{SaveOutput}
\usage{
SaveOutput(data)
}
\arguments{
\item{data}{List of data generates by the MultiExcelLoader oder Multivar function.}
}
\value{
Save CSV in Output Folder.
}
\description{
Save all Outputs from the data generated by MultiExcelLoader and Multivar as a CSV.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kroeger
}
