% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowMetaDataFilters.R
\name{ShowMetaDataFilters}
\alias{ShowMetaDataFilters}
\title{ShowMetaDataFilters}
\usage{
ShowMetaDataFilters(data)
}
\arguments{
\item{data}{List of data generates by the Multivar function.}
}
\value{
Returns a console output with all the Metadata values from witch data can be filtered.
}
\description{
Shows all the Metadata values from witch data can be filtered.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
\cr Comma numbers are rounded up.
}
\author{
Tim Kroeger
}
