% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesRichnessAsyncTabel.R
\name{SpeciesRichnessAsyncTabel}
\alias{SpeciesRichnessAsyncTabel}
\title{SpeciesRichnessAsyncTabel}
\usage{
SpeciesRichnessAsyncTabel(data, intervallBy = 100)
}
\arguments{
\item{data}{List of data generates by the Multivar function.}

\item{intervallBy}{Intervalls by to interpolate to.}
}
\value{
Returns the same List but with new added parameters.
}
\description{
Calculates the inverse species richness given by the SpeciesRichness function.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
\cr Comma numbers are rounded up.
}
\author{
Tim Kroeger
}
