% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRACETabel.R
\name{TRACETabel}
\alias{TRACETabel}
\title{TRACETabel}
\usage{
TRACETabel(data, TraceName = "UnknownTrace", intervallBy = 100)
}
\arguments{
\item{data}{List of data generates by the Multivar function.}

\item{TraceName}{Name of type of Trace data. For the search in the given list.}

\item{intervallBy}{Intervals by to mean to.}
}
\value{
Returns the same List but with new added parameters.
}
\description{
Calculates the TRACE as a Table.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
\cr Comma numbers are rounded up.
}
\author{
Tim Kroeger
}
