% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CutOutPionierPhase.R
\name{CutOutPionierPhase}
\alias{CutOutPionierPhase}
\title{CutOutPionierPhase}
\usage{
CutOutPionierPhase(data)
}
\arguments{
\item{data}{List of data generates by the Multivar function.}
}
\value{
Returns the same List but with new added parameters.
}
\description{
Cut out the Pionier Phase of the Lake data.
\cr The new stating value is the first Value that passes the T.test that for two samples with a P_value <=0.05.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
\cr Comma numbers are rounded up.
}
\author{
Tim Kroeger
}
