% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDicriptionCSV.R
\name{GetDiscriptionCSV}
\alias{GetDiscriptionCSV}
\title{GetDiscriptionCSV}
\usage{
GetDiscriptionCSV()
}
\value{
Save CSV in workspace.
}
\description{
Get a short Discription from all Cores and save them as csv.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kroeger
}
