% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivar.R
\name{Multivar}
\alias{Multivar}
\title{Multivar}
\usage{
Multivar(data, method = "bray", standardize = c("", "sqrt"))
}
\arguments{
\item{data}{List of data generates by the MultiExcelLoader function.}

\item{method}{Method for calculation Dissimilarity Indices for Community Ecologists}
}
\value{
Returns the same List but with new added parameters.
}
\description{
Calculates several parameters from deatom data.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
\cr Comma numbers are rounded up.
}
\author{
Tim Kroeger
}
