% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivar.R
\name{Multivar}
\alias{Multivar}
\title{Multivar}
\usage{
Multivar(
  data,
  method = "bray",
  standardize = c("", "sqrt"),
  percentFilterWeight = 0,
  allLoessSpans = 0.8,
  minimumRowsAfterFiltering = 0
)
}
\arguments{
\item{data}{List of data generates by the MultiExcelLoader function.}

\item{method}{Method for calculation Dissimilarity Indices for Community Ecologists.}

\item{standardize}{Method for data standardisation. Can be nothing "" ir sqaureroot transformation "sqrt".}

\item{percentFilterWeight}{Value how much percent a single species must relevant at minimum from the dataset.}

\item{allLoessSpans}{span value for all Loess calculations made by Multivar.}

\item{minimumRowsAfterFiltering}{minimum rows after filtering that must be}
}
\value{
Returns the same List but with new added parameters.
}
\description{
Calculates several parameters from deatom data.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
\cr Comma numbers are rounded up.
}
\author{
Tim Kroeger
}
